210513 좋아요 목록 mvc 완료

+++ src/main/java/com/example/demo/model/Like.java
@@ -14,7 +14,7 @@ public class Like {
     @Id
     @Column(name="like_no")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int likeNo;
+    private Long likeNo;
 
     // User pk
     @Column(name = "user_id")
